requirements.txt
future
PyGithub
requests
sqlalchemy>=0.8.0,<2.0.0
python-dateutil
alembic>=0.7.0
beautifulsoup4
rispy
Flask
Flask-RESTful
Flask-SQLAlchemy
marshmallow-sqlalchemy
flask-marshmallow
gunicorn
gevent
gsqlite3
git+https://gitlab.flux.utah.edu/searcch/giturlparse@76da9b6176a9b26f1ada12996c0dcfde6bd4b876
nltk
pdfminer.six
bibtexparser==0.6.2
git+https://github.com/LIAAD/yake.git@6a37efc67912195aebed562cd943bca8acc48f64
lxml
commonmark
Click
scikit-learn
git+https://github.com/boudinfl/pke.git@8f1d05dcc52041c9920ba0f9d5231fe6086d12c4
arxiv
atoma
pygit2==1.10.1
APScheduler
html5lib
